backend/requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]
python-multipart
python-dotenv
PyJWT